# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "旅游助手 API 文档",
        "title": "Travel Assistant API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/activity/route/calc": {
            "post": {
                "description": "计算两个活动之间的路线并保存，已存在的路线会从缓存直接返回",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "计算并保存活动路线",
                "parameters": [
                    {
                        "description": "计算活动路线请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CalcActivityRouteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "计算路线成功",
                        "schema": {
                            "$ref": "#/definitions/entity.ActivityRouteEntity"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.ActivityRouteEntity": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                },
                "duration": {
                    "type": "integer"
                },
                "endActivityId": {
                    "description": "终点活动id",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "startActivityId": {
                    "description": "起点活动id",
                    "type": "integer"
                },
                "tripId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "service.CalcActivityRouteRequest": {
            "type": "object",
            "required": [
                "endActivityId",
                "startActivityId"
            ],
            "properties": {
                "endActivityId": {
                    "type": "integer"
                },
                "forceRefresh": {
                    "type": "boolean"
                },
                "startActivityId": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}